add module to poll for git changes which trigger builds.
A set of project needs to be build whenever a source change is made to
any one of the branches composing the set.  The RepoStatus class
clones all build_specification git repositories into a tmp dir, and
tests them to see if the head has changed for any of the specified
branches.