add DependencyGraph
DependencyGraph process the prerequisites of a target build, and
indicates which builds have no prerequisites.  DependencyGraph is
notified of completed builds, and provides a list of subsequent builds
which have become unblocked.

Also removed "prerequisites" option, since it was always confusing.  A
separate scheduler will be made to build recursively instead of on
jenkins.