add Export functionality
When result_path is set, builders must copy their output from the
build_root to a directory that will be shared over nfs.  For now this
is accomplished with `cp -a`.  It may be necessary to select files
with globs in the future.

PKG_CONFIG settings were added to AutoBuilder, and were tested against
an m32 build.