Drop unnecessary delays
On the first access to some build info files, the project spec retries
reading the json files in case there is a network hiccup.  The file is
expected to not exist, so this operation was changed to not block.

Also fixed a program error which caused the revision specification
commandline generation to crash.