disable byt piglit tests except daily builds
For now, add a check to the piglit build to quit if the type is not
daily.  This allows us to enable piglit tests for that hardware,
whithout incurring 2hr delays in the build.

This change should be reverted when the build specification file can
explicitly set the type for prerequisites.