allow type-specific prerequisites
with the introduction of byt testers, test runs can take in excess of
2 hours.  This patch allows prerequisites to trigger builds only for
specific types, via the "only_for_type" attribute to the prequisite
tag.

the short-circuit of tests in piglit-test build.py was removed, as it
has been replaced by this mechanism.

Also, build_local.py was improved to skip hardware-specific builds.