drop vs-float-main-return from list of tests
In Oct 2014, a bunch of GLSL 1_10 tests were added that fail on most
of our platforms.  vs-float-main-return fails reliably on all the
platforms, except G45, where it fails intermittently.  This results in
confusing test results.

There is no value in running this test at all if it won't identify
regressions and produces only noise.