simplify branch configuration
projects which are omitted from a branch config will be held to the
default, but they will not trigger a build.  This should stop a lot of
the spurious building that occurs on developer branches.

Also, move the repo fetch command outside of the tight loop.  Repos
only have to update once a minute, not once per branch check.