build_bisect incorrectly built whatever rev was on disk
Unless the user specified the base_revision argument, build_bisect
would just take the revisions that are current in the workspace.  This
selection is arbitrary and dependent on the previous build.

A more sane choice is to choose mesa_master revisions as a starting
point, and override them if base_revision is provided.