implement generic bisect
add a global --retest_path to options, which targets only failing
tests from the old path to be executed.

use retest_path instead of extra_arg to target tests during bisection.
This option is more generic and can be used with other projects.

Factor out pass/fail detection from the bisector.  Bisector delegates
to the test class (so far, PiglitTest) to determine if a test failed
in bisection.

Enable pigilt-nir-test in bisection.