add byt-03
This machine is too slow to run daily tests.

However, it can be used for bisecting.  Setting the slave to go
offline when idle has the effect of giving it a lower priority than
the other byt slaves.  If a bisect is active during a daily build,
this machine will come online to take the bisect job.