change stable branches to use a fixed piglit
I have been spending useless hours updating config files to handle the
case where a new piglit test fails on the stable branch, and the
mainline commit is unknown.

Since new functionality is never backported onto stable, there is no
use in identifying which mesa commit provides the implementation for
the test.  Also, knowing that a test is fixed after the stable branch
point is not particularly helpful -- I still need to run a lengthy
bisect to find which commit could be possibly backported.

In the future, all stable branches should be created at a fixed piglit
revision.