[deqp] explicitly use python2
This is required for dEQP to run on Arch Linux. This path should be valid on Debian, see section 2.2 here: https://www.debian.org/doc/packaging-manuals/python-policy/ch-python.html

This may break other distros where `/usr/bin/python2` is not a valid pointer to python2, but I'm not aware of any.