Use 'flat' git repo caching
This converts the caching structure to a 'flat' model, where remotes are cached
in a single repo for each project. As a result, the poll_branches and
fetch_mirrors services have been adjusted to handle this new structure.

A new service, mesa_jenkins_monitor, was created to manage the poll_branches &
fetch_mirrors services (e.g. restarting them if they were updated or if
build_spec has been changed).